// src/store/mutation-types.ts
export enum MutationTypes {
  SET_INITIALIZED = 'SET_INITIALIZED',
  SET_LOADING = 'SET_LOADING',
  SET_SHOW_ALERT_MSG = 'SET_SHOW_ALERT_MSG',
  SET_ALERT_MSG = 'SET_ALERT_MSG',
  SET_ALERT_TYPE = 'SET_ALERT_TYPE',
  SET_API = 'SET_API',
  SET_EXTENSIONS = 'SET_EXTENSIONS',
  SET_CURRENT_NETWORK_STATUS = 'SET_CURRENT_NETWORK_STATUS',
  SET_CURRENT_NETWORK_IDX = 'SET_CURRENT_NETWORK_IDX',
  SET_CURRENT_ACCOUNT_IDX = 'SET_CURRENT_ACCOUNT_IDX',
  SET_CURRENT_CUSTOM_ENDPOINT = 'SET_CURRENT_CUSTOM_ENDPOINT',
  SET_THEME = 'SET_THEME'
}